refactor(cli): convert to typescript sources
