+++ cli/bin/mew.js
@@ -10,5 +10,13 @@
  * - mew token create     - Create a test token
  */
 
-require('../src/index.js');
+try {
+  require('../dist/index.js');
+} catch (error) {
+  if (error && error.code === 'MODULE_NOT_FOUND') {
+    console.error('The MEW CLI has not been built yet. Please run "npm run build" inside the cli package.');
+    process.exit(1);
+  }
+  throw error;
+}
