+++ cli/src/commands/gateway.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 const { Command } = require('commander');
 const WebSocket = require('ws');
 const express = require('express');
