+++ cli/src/commands/space.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 const { Command } = require('commander');
 const fs = require('fs');
 const path = require('path');
