+++ cli/src/config/keyBindings.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /**
  * Key Bindings Configuration for MEW CLI
  *
