+++ cli/src/index.ts
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+// @ts-nocheck
 
 const { program } = require('commander');
 const fs = require('fs');
