+++ cli/src/ui/keyMatchers.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /**
  * Key Matchers for Terminal Input
  *
