+++ cli/src/ui/utils/textUtils.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /**
  * Text Utilities for Unicode and Terminal Handling
  *
