+++ cli/src/utils/banner.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /**
  * MEW Protocol Banner
  *
