+++ cli/src/utils/interactive-ui.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /**
  * Interactive Terminal UI
  *
