+++ cli/src/utils/message-input.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /**
  * Message Input Utility
  *
