+++ cli/src/utils/participant-resolver.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 /**
  * Participant Resolution Logic
  *
