+++ tsconfig.json
@@ -6,6 +6,7 @@
     { "path": "./sdk/typescript-sdk/client" },
     { "path": "./sdk/typescript-sdk/participant" },
     { "path": "./sdk/typescript-sdk/agent" },
-    { "path": "./bridge" }
+    { "path": "./bridge" },
+    { "path": "./cli" }
   ]
 }
